TMDb API - Top rated movies and others
Simple App that uses the TMDb API through Retrofit. It also uses CardView and RecyclerView to display the movies with a nice card interface, Glide to load images from URLs, Palette to get the most common colors in pictures and Picasso to convert to Bitmap images from URLs.
I based this project in some tutorials that I found on http://androidhive.info